requirements.txt
Django==3.2
django-ckeditor==6.2.0
djangorestframework==3.12.4
django-redis==5.2.0
redis==3.5.3
googletrans==4.0.0-rc1
pytest==7.4.0
flake8==6.0.0
django-redis
black